+++ src/adc/filter/_metadata_from_name.py
@@ -8,16 +8,26 @@ from seppl.io import Filter
 from wai.logging import LOGGING_WARNING
 
 
+SOURCE_NAME = "name"
+SOURCE_PATH = "path"
+SOURCES = [
+    SOURCE_NAME,
+    SOURCE_PATH,
+]
+
+
 class MetaDataFromName(Filter):
     """
     Extracts a sub-string from the audio name and stores them in the meta-data.
     """
 
-    def __init__(self, regexp: str = None, metadata_key: str = None,
+    def __init__(self, source: str = None, regexp: str = None, metadata_key: str = None,
                  logger_name: str = None, logging_level: str = LOGGING_WARNING):
         """
         Initializes the filter.
 
+        :param source: where to apply the regexp to, default is name
+        :type source: str
         :param regexp: the regular expression to apply to the name (1st group gets used as label)
         :type regexp: str
         :param metadata_key: the metadata key to store the extracted substring under
@@ -28,6 +38,7 @@ class MetaDataFromName(Filter):
         :type logging_level: str
         """
         super().__init__(logger_name=logger_name, logging_level=logging_level)
+        self.source = source
         self.regexp = regexp
         self.metadata_key = metadata_key
 
@@ -75,6 +86,7 @@ class MetaDataFromName(Filter):
         :rtype: argparse.ArgumentParser
         """
         parser = super()._create_argparser()
+        parser.add_argument("-s", "--source", choices=SOURCES, help="The string to apply the regular expression to for extracting the meta-data.", default=SOURCE_NAME, required=False)
         parser.add_argument("-r", "--regexp", type=str, help="The regular expression apply to the audio name, with the 1st group being used as the meta-data value.", default=None, required=False)
         parser.add_argument("-k", "--metadata_key", type=str, help="The key in the meta-data to store the extracted sub-string under.", default=None, required=False)
         return parser
@@ -87,6 +99,7 @@ class MetaDataFromName(Filter):
         :type ns: argparse.Namespace
         """
         super()._apply_args(ns)
+        self.source = ns.source
         self.regexp = ns.regexp
         self.metadata_key = ns.metadata_key
 
@@ -95,6 +108,8 @@ class MetaDataFromName(Filter):
         Initializes the processing, e.g., for opening files or databases.
         """
         super().initialize()
+        if self.source is None:
+            self.source = SOURCE_NAME
         if self.metadata_key is None:
             raise Exception("No meta-data key provided!")
 
@@ -105,24 +120,35 @@ class MetaDataFromName(Filter):
         :param data: the record to update
         """
         try:
-            name = data.audio_name
-            if name is None:
-                self.logger().warning("No audio name available: %s" % str(data))
+            # get string for regexp
+            if self.source == SOURCE_NAME:
+                source = data.audio_name
+            elif self.source == SOURCE_PATH:
+                source = data.source
+            else:
+                raise Exception("Unhandled source: %s" % self.source)
+            if source is None:
+                self.logger().warning("No %s available to apply regexp to: %s" % (self.source, str(data)))
                 return data
-            m = re.search(self.regexp, name)
+
+            # apply regexp
+            m = re.search(self.regexp, source)
             if m is None:
                 return data
             value = m.group(1)
+
+            # set meta-data
             result = copy.deepcopy(data)
             if not result.has_metadata():
                 meta = dict()
             else:
                 meta = result.get_metadata()
             meta[self.metadata_key] = value
+            self.logger().info("%s = %s" % (self.metadata_key, value))
             result.set_metadata(meta)
             return result
         except:
-            self.logger().exception("Failed to extract meta-data value from: %s" % data.audio_name)
+            self.logger().exception("Failed to extract meta-data value from: %s" % str(data))
             return data
 
     def _do_process(self, data):