+++ CHANGES.rst
@@ -8,6 +8,7 @@ Changelog
 - centralized comparison code in `adc.api._comparison` (available via `adc.api`)
 - the `tee` meta-filter can now forward or drop the incoming data based on a meta-data evaluation
 - added `sub-process` filter for processing data with sub-flow of filters, can be conditional based on meta-data evaluation
+- the `metadata-from-name` filter can work on the path now as well (must be present)
 
 
 0.0.4 (2025-07-15)